Convert photo gallery to new modal system
