+++ frontend/src/app/@modal/sign-in/page.tsx
@@ -1,5 +1,6 @@
-import {CloseModalButton, SignInForm} from '@/app/@modal/sign-in/page.client'
+import {SignInForm} from '@/app/@modal/sign-in/page.client'
 import {Modal, ModalActions} from '@/components/Modal'
+import {CloseModalButton} from '@/components/CloseModalButton'
 
 export default function SignInModal () {
     return (
