+++ frontend/src/app/@modal/sign-up/page.tsx
@@ -1,6 +1,6 @@
-import {CloseModalButton} from '@/app/@modal/sign-in/page.client'
 import {Modal, ModalActions} from '@/components/Modal'
 import {SignUpForm} from '@/app/@modal/sign-up/page.client'
+import {CloseModalButton} from '@/components/CloseModalButton'
 
 export default function SignUpModal () {
     return (
